+++ src/briefy/leica/models/customer/workflows/__init__.py
@@ -1,0 +1,4 @@
+"""Package containing workflows for Customer."""
+from .address import BillingAddressWorkflow  # noQA
+from .contact import ContactWorkflow  # noQA
+from .customer import CustomerWorkflow  # noQA