+++ src/briefy/leica/models/professional/workflows/professional.py
@@ -4,12 +4,9 @@ from briefy.common.vocabularies.roles import LocalRolesChoices as LR
 from briefy.common.workflow import WorkflowState as WS
 from briefy.common.workflow import BriefyWorkflow
 from briefy.common.workflow import Permission
-from briefy.leica.utils.user import create_rolleiflex_user
-
-import logging
-
-
-logger = logging.getLogger(__name__)
+from briefy.leica.utils.user import activate_or_create_user
+from briefy.leica.utils.user import activate_user
+from briefy.leica.utils.user import inactivate_user
 
 
 class ProfessionalWorkflow(BriefyWorkflow):
@@ -55,7 +52,7 @@ class ProfessionalWorkflow(BriefyWorkflow):
     )
 
     deleted = WS(
-        'deleted', 'Deleted Skill',
+        'deleted', 'Deleted professional',
         'Professional was deleted from the platform.'
     )
 
@@ -68,7 +65,7 @@ class ProfessionalWorkflow(BriefyWorkflow):
     @validation.transition(deleted, 'can_delete')
     @pending.transition(deleted, 'can_delete')
     def delete(self):
-        """Delete a professional."""
+        """Delete a professional from our platform."""
         pass
 
     @created.transition(pending, 'can_submit')
@@ -86,7 +83,7 @@ class ProfessionalWorkflow(BriefyWorkflow):
     def approve(self):
         """Quality approval of a professional application."""
         groups = ('g:professionals',)
-        create_rolleiflex_user(self.document, groups=groups)
+        activate_or_create_user(self.document, groups=groups)
 
     @validation.transition(trial, 'can_validate')
     def validate(self):
@@ -97,13 +94,16 @@ class ProfessionalWorkflow(BriefyWorkflow):
     @inactive.transition(active, 'can_activate')
     def activate(self):
         """Activate a professional."""
-        pass
+        professional = self.document
+        # Activate on Rolleiflex only if current state is inactive.
+        if professional.state == 'inactive':
+            activate_user(professional)
 
     @active.transition(inactive, 'can_inactivate')
     @trial.transition(inactive, 'can_inactivate')
     def inactivate(self):
         """Inactivate a professional."""
-        pass
+        inactivate_user(self.document)
 
     @active.transition(active, 'can_assign', required_fields=('pools_ids',))
     @trial.transition(trial, 'can_assign', required_fields=('pools_ids',))
@@ -128,7 +128,7 @@ class ProfessionalWorkflow(BriefyWorkflow):
         """Validate if user can submit this professional for QA approval."""
         return True
 
-    @Permission(groups=[LR['owner'], G['qa'], ])
+    @Permission(groups=[LR['owner'], G['qa'], G['scout'], ])
     def can_quality_reject(self):
         """Validate if user can reject this professional application."""
         return True
@@ -161,100 +161,4 @@ class ProfessionalWorkflow(BriefyWorkflow):
     @Permission(groups=[G['system'], G['pm'], G['scout']])
     def can_assign(self):
         """Validate if user can assign this professional to a pool."""
-        return True
-
-
-class LocationWorkflow(BriefyWorkflow):
-    """Workflow for a Working Location."""
-
-    entity = 'workinglocation'
-    initial_state = 'created'
-
-    # States
-    created = WS(
-        'created', 'Created',
-        'Inserted into the database.'
-    )
-
-    active = WS(
-        'active', 'Active working location',
-        'Professional can be hired at this working location.'
-    )
-
-    inactive = WS(
-        'inactive', 'Inactive working location',
-        'Professional cannot be hired at this working location.'
-    )
-
-    deleted = WS(
-        'deleted', 'Deleted working location',
-        'Working location was deleted from the platform.'
-    )
-
-    # Transitions:
-    @created.transition(active, 'can_submit')
-    @inactive.transition(active, 'can_certify')
-    def submit(self):
-        """Transition a working locaiton from created or inactive to active."""
-        pass
-
-    @active.transition(inactive, 'can_inactivate')
-    def inactivate(self):
-        """Inactivate a working location."""
-        pass
-
-    @active.transition(deleted, 'can_delete')
-    @inactive.transition(deleted, 'can_delete')
-    def delete(self):
-        """Delete a working location."""
-        pass
-
-    # Permissions:
-    @Permission(groups=[G['professionals'], ])
-    def can_submit(self):
-        """Validate if user can submit a profile."""
-        return True
-
-    @Permission(groups=[LR['owner'], G['scout'], ])
-    def can_activate(self):
-        """Validate if user can activate this working location."""
-        return True
-
-    @Permission(groups=[LR['owner'], G['scout'], ])
-    def can_inactivate(self):
-        """Validate if user can inactivate this working location."""
-        return True
-
-    @Permission(groups=[LR['owner'], G['scout'], ])
-    def can_delete(self):
-        """Validate if user can delete this working location."""
-        return True
-
-
-class LinkWorkflow(BriefyWorkflow):
-    """Workflow for a link."""
-
-    entity = 'link'
-    initial_state = 'created'
-
-    # States
-    created = WS(
-        'created', 'Created',
-        'Inserted into the database.'
-    )
-
-    deleted = WS(
-        'deleted', 'Deleted link',
-        'Link was deleted from the platform.'
-    )
-
-    # Transitions:
-    @created.transition(deleted, 'can_delete')
-    def delete(self):
-        """Delete a link."""
-        pass
-
-    @Permission(groups=[LR['owner'], G['scout'], ])
-    def can_delete(self):
-        """Validate if user can delete this link."""
         return True