+++ src/briefy/leica/models/user/workflows/__init__.py
@@ -1,0 +1,4 @@
+"""Package containing workflows for User Profiles."""
+from .base import UserProfileWorkflow  # noQA
+from .customer import CustomerUserProfileWorkflow  # noQA
+from .internal import BriefyUserProfileWorkflow  # noQA