+++ src/briefy/leica/models/user/workflows/base.py
@@ -1,14 +1,8 @@
-"""User profile workflow."""
-from briefy.common.vocabularies.roles import Groups as G
+"""Base User profile workflow."""
 from briefy.common.workflow import WorkflowState as WS
 from briefy.common.workflow import BriefyWorkflow
-from briefy.common.workflow import Permission
-from briefy.leica.utils.user import create_rolleiflex_user
-
-import logging
-
-
-logger = logging.getLogger(__name__)
+from briefy.leica.utils.user import activate_or_create_user
+from briefy.leica.utils.user import inactivate_user
 
 
 class UserProfileWorkflow(BriefyWorkflow):
@@ -44,32 +38,10 @@ class UserProfileWorkflow(BriefyWorkflow):
             groups = ('g:customers', )
         elif profile.type == 'briefyuserprofile':
             groups = ('g:briefy',)
-        create_rolleiflex_user(self.document, groups=groups)
-
-    @Permission(groups=[G['system'], G['pm'], G['scout'], G['bizdev'], G['finance']])
-    def can_activate(self):
-        """Validate if user can activate this user profile."""
-        return True
+        activate_or_create_user(self.document, groups=groups)
 
     # Transitions:
     @active.transition(inactive, 'can_inactivate')
     def inactivate(self):
         """Inactivate the UserProfile."""
-        pass
-
-    @Permission(groups=[G['system'], G['pm'], G['scout'], G['bizdev'], G['finance']])
-    def can_inactivate(self):
-        """Validate if user can inactivate this user profile."""
-        return True
-
-
-class CustomerUserProfileWorkflow(UserProfileWorkflow):
-    """Workflow for a Customer User profile."""
-
-    entity = 'customeruserprofile'
-
-
-class BriefyUserProfileWorkflow(UserProfileWorkflow):
-    """Workflow for a Briefy user profile."""
-
-    entity = 'briefyuserprofile'
+        inactivate_user(self.document)