+++ src/briefy/leica/utils/user.py
@@ -1,12 +1,13 @@
 """Utils to query user info."""
-from briefy.leica import logger
-from briefy.leica.config import API_BASE
-from briefy.leica.models.mixins import get_public_user_info
-from pyramid.httpexceptions import HTTPBadRequest
+from briefy.common.utilities.interfaces import IUserProfileQuery
+#  from briefy.leica import logger
+from briefy.leica.utils.rolleiflex import create_user
+from briefy.leica.utils.rolleiflex import get_user
+from briefy.leica.utils.rolleiflex import transition_user
+from typing import Sequence
+from zope.component import getUtility
 
-import json
 import random
-import requests
 import string
 
 
@@ -15,13 +16,8 @@ def add_user_info_to_state_history(state_history):
 
     :param state_history: list of workflow state history.
     """
-    for item in state_history:
-        user = item.get('actor', None)
-        if isinstance(user, str):
-            user_id = user  # first call where actor is a UUID string
-            new_actor = get_public_user_info(user_id)
-            if new_actor:
-                item['actor'] = new_actor
+    profile_service = getUtility(IUserProfileQuery)
+    return profile_service.update_wf_history(state_history)
 
 
 def password_generator(size=8, chars=string.ascii_uppercase + string.digits):
@@ -32,40 +28,58 @@ def password_generator(size=8, chars=string.ascii_uppercase + string.digits):
 
 def create_rolleiflex_user(profile, groups=()):
     """Create a new Rolleiflex user from a UserProfile."""
-    # add a rolleiflex user
-    url = API_BASE + '/users'
-    groups = [{'name': value for value in groups}]
     initial_password = password_generator()
-    payload = dict(
-        id=str(profile.id),
-        email=profile.email,
-        first_name=profile.first_name,
-        last_name=profile.last_name,
-        password=initial_password,
-        locale='en_GB',
-        groups=groups,
-    )
+    return create_user(profile, initial_password, groups)
 
-    headers = {
-        'x-locale': 'en_GB',
-        'content-type': 'application/json'
-    }
 
-    response = requests.request('POST', url, data=json.dumps(payload), headers=headers)
-    if response.status_code in (200, 201):
-        logger.info('Success user creation. Email: {email}.'.format(email=profile.email))
-        # We put this here to be serialized to sqs
-        profile.initial_password = initial_password
-        return response.json()
+def activate_or_create_user(profile: 'UserProfile', groups: Sequence=()) -> bool:
+    """Create an User on Rolleiflex or activate an existing User.
+
+    :param profile: UserProfile to be synced to Rolleiflex.
+    :param groups: Sequence of groups to be used on User creation
+    :return: Status of this action.
+    """
+    user_id = profile.id
+    user = get_user(user_id)
+    if user:
+        status = transition_user(user_id, 'activate')
     else:
-        try:
-            result = response.json()
-        except Exception as exc:
-            msg = 'Failure to create user on Rolleiflex. Exception: {exc}'.format(exc=exc)
-        else:
-            default_msg = 'Error message not found. Failure to create user on Rolleiflex.'
-            msg = result.get('message', default_msg)
+        status = True if create_rolleiflex_user(profile, groups) else False
+    return status
 
-        logger.error(msg)
-        # TODO: improve exception handling
-        raise HTTPBadRequest(msg)
+
+def _transition_profile(profile: 'UserProfile', transition: str) -> bool:
+    """Transition an User connected to an UserProfile.
+
+    :param profile: UserProfile to be transitioned on Rolleiflex.
+    :param transition: Transition to be executed.
+    :return: Status of this action
+    """
+    user_id = profile.id
+    user = get_user(user_id)
+    if user:
+        status = transition_user(user_id, transition)
+    else:
+        # TODO: Raise exception here
+        status = False
+    return status
+
+
+def activate_user(profile: 'UserProfile') -> bool:
+    """Activate an User connected to an UserProfile.
+
+    :param profile: UserProfile to be transitioned on Rolleiflex.
+    :return: Status of this action
+    """
+    transition = 'activate'
+    return _transition_profile(profile, transition)
+
+
+def inactivate_user(profile: 'UserProfile') -> bool:
+    """Inactivate an User connected to an UserProfile.
+
+    :param profile: UserProfile to be transitioned on Rolleiflex.
+    :return: Status of this action
+    """
+    transition = 'inactivate'
+    return _transition_profile(profile, transition)